Update README.md
This project is focused on building a regression model to predict IMDb movie ratings based on various attributes such as genre, director, cast, runtime, and more. The goal is to analyze how different features influence movie ratings and to develop a predictive system that can estimate ratings for new or unseen movies.